requirements.txt
langchain>=0.2.0,<0.4.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-nvidia-ai-endpoints>=0.3.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
torch>=2.0.0
gradio>=5.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0